:art: Refactor: 마이페이지 폴더 추가
